requirements.txt
PyYAML~=5.3.1
SQLAlchemy~=1.3.20
discord.py~=1.6.0
cryptography~=3.1.1
psutil~=5.7.2
watchdog~=0.10.3